backend/requirements.txt
django
django-ninja
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-ninja-jwt
django-cors-headers
cryptography
llama-cpp-python
PyPDF2
python-docx
chardet
requests
aiohttp
celery
redis